+++ src/ui_callbacks.h
@@ -21,6 +21,4 @@ unsigned int io_seproxyhal_touch_eth2_address_ok(const bagl_element_t *e);
 unsigned int io_seproxyhal_touch_privacy_ok(const bagl_element_t *e);
 unsigned int io_seproxyhal_touch_privacy_cancel(const bagl_element_t *e);
 
-void ui_warning_contract_data(void);
-
 void io_seproxyhal_send_status(uint32_t sw);