+++ src_features/getAppConfiguration/cmd_getAppConfiguration.c
@@ -12,7 +12,8 @@ void handleGetAppConfiguration(uint8_t p1,
     UNUSED(workBuffer);
     UNUSED(dataLength);
     UNUSED(flags);
-    G_io_apdu_buffer[0] = APP_FLAG_EXTERNAL_TOKEN_NEEDED;
+    G_io_apdu_buffer[0] = (N_storage.dataAllowed ? APP_FLAG_DATA_ALLOWED : 0x00);
+    G_io_apdu_buffer[0] |= APP_FLAG_EXTERNAL_TOKEN_NEEDED;
     G_io_apdu_buffer[1] = MAJOR_VERSION;
     G_io_apdu_buffer[2] = MINOR_VERSION;
     G_io_apdu_buffer[3] = PATCH_VERSION;